SCandidate-Institute Module

+++ Backend/components/scandidate/organization-onboard/orgOnboardController.js
@@ -1,17 +1,19 @@
 const AppError = require("./../../../helpers/appError");
 const colors = require("./../../../helpers/colors");
 const orgnisationDAL = require("./orgOnboardDAL");
+const orgnisationValidator = require("./orgOnboardValidator");
 const mongoose = require("mongoose");
 
 //Add Organisation
 module.exports.onboardOrganisationMethod = async function (req, res, next) {
   const data = req.body;
-      try {
-      let organisationData = await orgnisationDAL.onboardOrganisation(data);
-      return res.status(200).json({ status: "SUCCESS", data: organisationData });
-    } catch (err) {
-      return next(new AppError(err, 400));
-    }
+  try {
+    await orgnisationValidator.organisationCreationSchema.validateAsync(data);
+    let organisationData = await orgnisationDAL.onboardOrganisation(data);
+    return res.status(200).json({ status: "SUCCESS", data: organisationData });
+  } catch (err) {
+    return next(new AppError(err, 400));
+  }
 };
 
 
@@ -30,9 +32,10 @@ module.exports.getAllMethod = async function (req, res, next) {
 
 //Get Organisation By Id
 module.exports.getOrganisationByIdMethod = async function (req, res, next) {
-  const data = { _id: mongoose.Types.ObjectId(req.params.organisationId) };
   try {
+    const data = { _id: mongoose.Types.ObjectId(req.params.organisationId) };
     let organisationData = await orgnisationDAL.getOrganisationById(data);
+    if (!organisationData) return next(new AppError("Organisation does not exists!", 404));
     return res.status(200).json({
       status: "SUCCESS",
       message: null,
@@ -43,13 +46,16 @@ module.exports.getOrganisationByIdMethod = async function (req, res, next) {
   }
 };
 
-
 //Update Organisation
 module.exports.updateOrganisationMethod = async function (req, res, next) {
   const data = req.body;
-  data._id = mongoose.Types.ObjectId(req.params.organisationId);
   try {
-    let organisationData = await orgnisationDAL.updateOrganisation(data);
+    let result = await orgnisationValidator.organisationUpdationSchema.validateAsync(data);
+    data._id = mongoose.Types.ObjectId(req.params.organisationId); 
+    const organisationExsits = await orgnisationDAL.getOrganisationById(data);
+    if (!organisationExsits) return next(new AppError("user does not exists!", 404));
+    result._id = mongoose.Types.ObjectId(req.params.organisationId); 
+    let organisationData = await orgnisationDAL.updateOrganisation(result);
     return res.status(200).json({
       status: "SUCCESS",
       message: "Organisation Data has been updated successfully",
@@ -60,12 +66,14 @@ module.exports.updateOrganisationMethod = async function (req, res, next) {
   }
 };
 
-//Delete Organisation
-module.exports.deleteOrganisationMethod = async function (req, res) {
-  const data = { _id: mongoose.Types.ObjectId(req.params.organisationId) };
-  try {
-    let organisationData = await orgnisationDAL.deleteOrganisation(data);
 
+//Delete Organisation
+module.exports.deleteOrganisationMethod = async function (req, res,next) {
+  try {
+    const data = { _id: mongoose.Types.ObjectId(req.params.organisationId) };
+    const organisationExsits = await orgnisationDAL.getOrganisationById(data);
+    if (!organisationExsits) return next(new AppError("Organisation does not exists!", 404));
+    const organisationData = await orgnisationDAL.deleteOrganisation(data);
     return res.status(200).json({
       status: "SUCCESS",
       message: "Organisation has been deleted successfully !",
